Cleaning up code for merge, code runs, but ipopt crashes
