+++ examples/Cassie/BUILD.bazel
@@ -351,17 +351,68 @@ cc_binary(
 cc_binary(
     name = "centroidal_mpc_tracking_controller",
     srcs = ["centroidal_mpc_tracking_controller.cc"],
+    data = ["//examples/Cassie/kinematic_centroidal_planner:gains"],
     deps = [
         ":cassie_urdf",
         ":cassie_utils",
         "//examples/Cassie/osc_jump",
-        "//examples/Cassie/kinematic_centroidal_planner:centroidal_mpc",
+        "//examples/Cassie/kinematic_centroidal_planner:centroidal_tracking_controller",
         "//lcm:trajectory_saver",
         "//systems/trajectory_optimization:lcm_trajectory_systems",
         "//multibody:utils",
         "//systems:robot_lcm_systems",
         "//systems:system_utils",
         "//systems/primitives:trajectory_passthrough",
+        "//systems/framework:lcm_driven_loop",
+        "//systems/primitives",
+        "@drake//:drake_shared_library",
+        "@gflags",
+    ],
+)
+
+cc_binary(
+    name = "centroidal_mpc_reference",
+    srcs = ["centroidal_mpc_reference.cc"],
+    data = ["//examples/Cassie/kinematic_centroidal_planner:kinematic_centroidal_mpc_gains.yaml",
+            "//examples/Cassie/kinematic_centroidal_planner/gaits:gaits",
+            "//examples/Cassie/kinematic_centroidal_planner/motions:motions"],
+    deps = [
+        ":cassie_urdf",
+        ":cassie_utils",
+        "//examples/Cassie/osc:osc",
+        "//examples/Cassie/systems:cassie_out_to_radio",
+        "//examples/Cassie/kinematic_centroidal_planner:centroidal_tracking_controller",
+        "//lcm:trajectory_saver",
+        "//systems/trajectory_optimization:lcm_trajectory_systems",
+        "//systems/trajectory_optimization/kinematic_centroidal_planner:kinematic_centroidal_planner",
+        "//systems/controllers/kinematic_centroidal_mpc:kinematic_centroidal_mpc",
+        "//multibody:utils",
+        "//systems:robot_lcm_systems",
+        "//systems:system_utils",
+        "//systems/primitives:trajectory_passthrough",
+        "//systems/framework:lcm_driven_loop",
+        "//systems/primitives",
+        "@drake//:drake_shared_library",
+        "@gflags",
+    ],
+)
+
+cc_binary(
+    name = "run_mpc_visualization",
+    srcs = ["run_mpc_visualization.cc"],
+    deps = [
+        ":cassie_urdf",
+        ":cassie_utils",
+        "//examples/Cassie/osc:osc",
+        "//examples/Cassie/systems:cassie_out_to_radio",
+        "//examples/Cassie/kinematic_centroidal_planner:centroidal_tracking_controller",
+        "//lcm:trajectory_saver",
+        "//systems/controllers/kinematic_centroidal_mpc:kinematic_centroidal_visualizer",
+        "//multibody:utils",
+        "//systems:robot_lcm_systems",
+        "//systems:system_utils",
+        "//systems/primitives:trajectory_passthrough",
+        "//systems/trajectory_optimization:lcm_trajectory_systems",
         "//systems/framework:lcm_driven_loop",
         "//systems/primitives",
         "@drake//:drake_shared_library",
