+++ examples/Cassie/kinematic_centroidal_planner/gaits/right_step.yaml
@@ -1,0 +1,8 @@
+gait_pattern:
+  - start_phase: 0.0
+    end_phase: 0.7
+    contact_status: [ true, true, false, false ]
+  - start_phase: 0.7
+    end_phase: 1.0
+    contact_status: [ true, true, true, true ]
+period: 0.5
