--- examples/Cassie/kinematic_centroidal_planner/gaits/step.yaml
@@ -1,8 +1,0 @@
-gait_pattern:
-  - start_phase: 0.0
-    end_phase: 0.9
-    contact_status: [true, true, false, false]
-  - start_phase: 0.9
-    end_phase: 1.0
-    contact_status: [true, true, true, true]
-period: 0.5
