+++ examples/Cassie/kinematic_centroidal_planner/motions/motion_walk.yaml
@@ -1,0 +1,11 @@
+n_knot_points: 45
+target_com_height: 0.8
+stance_width: 0.25
+swing_foot_minimum_height: 0.03
+duration_scaling:
+  [ 1.0, 1.0, 1.0 ]
+gait_sequence: [ 'stand', 'walk', 'walk' ]
+com_vel_values:
+  [ 0, 0, 0,
+    0, 0, 0,
+    0, 0, 0 ]
