+++ systems/trajectory_optimization/BUILD.bazel
@@ -55,6 +55,7 @@ cc_library(
     ],
     deps = [
         "//lcm:lcm_trajectory_saver",
+        "//common:find_resource",
         "@drake//:drake_shared_library",
     ],
 )
