+++ systems/trajectory_optimization/kinematic_centroidal_planner/BUILD.bazel
@@ -1,0 +1,52 @@
+package(default_visibility = ["//visibility:public"])
+
+cc_library(
+    name = "kinematic_centroidal_planner",
+    srcs = [],
+    deps = [
+        ":kinematic_centroidal_solver",
+        ":kcmpc_gains",
+        ":kinematic_centroidal_reference_generator",
+    ],
+)
+
+cc_library(
+    name = "kinematic_centroidal_solver",
+    srcs = ["kinematic_centroidal_solver.cc",
+            "kinematic_centroidal_constraints.cc"],
+    hdrs = ["kinematic_centroidal_solver.h",
+            "kinematic_centroidal_constraints.h"],
+    deps = [
+        "//solvers:constraints",
+        "//multibody/kinematic:kinematic",
+        "//multibody/kinematic:constraints",
+        "//multibody:multipose_visualizer",
+        ":kcmpc_gains",
+        "//lcm:lcm_trajectory_saver",
+        "@drake//:drake_shared_library",
+    ],
+)
+
+cc_library(
+    name = "kcmpc_gains",
+    hdrs = ["kinematic_centroidal_gains.h"],
+    deps = [
+        "@drake//:drake_shared_library",
+    ],
+)
+
+cc_library(
+    name = "kinematic_centroidal_reference_generator",
+    srcs = ["kcmpc_reference_generator.cc",
+            "reference_generation_utils.cc",
+            "gait.cc"],
+    hdrs = ["kcmpc_reference_generator.h",
+            "reference_generation_utils.h",
+            "gait.h"],
+    deps = [
+        "//common",
+        "//examples/Cassie:cassie_utils",
+        "//multibody/kinematic:kinematic",
+        "@drake//:drake_shared_library",
+    ],
+)
