+++ systems/trajectory_optimization/kinematic_centroidal_planner/gait.h
@@ -1,16 +1,19 @@
 #pragma once
 
 #include <drake/common/trajectories/piecewise_polynomial.h>
-#include "drake/common/yaml/yaml_read_archive.h"
+
 #include "yaml-cpp/yaml.h"
+
+#include "drake/common/yaml/yaml_read_archive.h"
 
 /*!
  * @brief Struct for a given mode as part of a mode sequence
  */
-struct Mode{
+struct Mode {
   double start_phase;
   double end_phase;
-  drake::VectorX<bool> contact_status; /// Vector describing which contacts are active
+  /// Vector describing which contacts are active
+  drake::VectorX<bool> contact_status;
 
   template <typename Archive>
   void Serialize(Archive* a) {
@@ -21,16 +24,19 @@ struct Mode{
 };
 
 /*!
- * @brief Struct for defining a gait consisting of a vector of Modes, and a period for the whole gait.
+ * @brief Struct for defining a gait consisting of a vector of Modes, and a
+ * period for the whole gait.
  */
-struct Gait{
-/*!
- * @brief converts the gait into a trajectory from current time to end time
- * @param current_time
- * @param end_time
- * @return trajectory where value at time t converted to a bool is a vector of which feet are in stance
- */
-  drake::trajectories::PiecewisePolynomial<double> ToTrajectory(double current_time, double end_time) const;
+struct Gait {
+  /*!
+   * @brief converts the gait into a trajectory from current time to end time
+   * @param current_time
+   * @param end_time
+   * @return trajectory where value at time t converted to a bool is a vector of
+   * which feet are in stance
+   */
+  drake::trajectories::PiecewisePolynomial<double> ToTrajectory(
+      double current_time, double end_time) const;
 
   /*!
    * @brief Checks to make sure the gait is valid, Asserts if not valid
@@ -50,7 +56,8 @@ struct Gait{
   int GetCurrentMode(double time_now) const;
 
   double period;
-  std::vector<Mode> gait_pattern; /// Vector of modes. Start phase of the first mode must be 0, end phase of the last mode
+  /// Vector of modes. Start phase of the first mode must be
+  /// 0, end phase of the last mode
   /// must be 1, and no time gaps between start and end of sequential modes
-
+  std::vector<Mode> gait_pattern;
 };
