+++ systems/trajectory_optimization/kinematic_centroidal_planner/kcmpc_reference_generator.h
@@ -1,32 +1,35 @@
 #pragma once
 
-#include <drake/multibody/plant/multibody_plant.h>
 #include <drake/common/trajectories/piecewise_polynomial.h>
+#include <drake/multibody/plant/multibody_plant.h>
+
 #include "multibody/kinematic/world_point_evaluator.h"
-#include "systems/controllers/kinematic_centroidal_mpc/gait.h"
+#include "systems/trajectory_optimization/kinematic_centroidal_planner/gait.h"
 
 /*!
- * @brief struct consisting of a collection of times and samples used for passing inputs to reference generator
+ * @brief struct consisting of a collection of times and samples used for
+ * passing inputs to reference generator
  * @tparam T
  */
 template <typename T>
-struct KnotPoints{
+struct KnotPoints {
   std::vector<double> times;
   std::vector<T> samples;
 };
 
-class KcmpcReferenceGenerator{
+class KcmpcReferenceGenerator {
  public:
-
   /*!
    * @brief constructor for reference generator
    * @param plant
    * @param nominal_stand nominal stand of size [nq]
    * @param contacts Vector of world point evalator describing contacts
    */
-  KcmpcReferenceGenerator(const drake::multibody::MultibodyPlant<double>& plant,
-                          const Eigen::VectorXd& nominal_stand,
-                          const std::vector<dairlib::multibody::WorldPointEvaluator<double>>& contacts);
+  KcmpcReferenceGenerator(
+      const drake::multibody::MultibodyPlant<double>& plant,
+      drake::systems::Context<double>* context,
+      const std::vector<dairlib::multibody::WorldPointEvaluator<double>>&
+          contacts);
 
   /*!
    * @brief specify com velocity at specific times
@@ -41,15 +44,26 @@ class KcmpcReferenceGenerator{
   void SetGaitSequence(const KnotPoints<Gait>& gait_knot_points);
 
   /*!
-   * @brief construct references assuming initial com is the from the nominal stand
+   * @brief return gait sequence
+   * @param gait_knot_points
    */
-  void Build();
+  static std::vector<double> GenerateTimePoints(
+      const std::vector<double>& duration_scaling,
+      std::vector<Gait> gait_knot_points);
 
   /*!
-   * @brief Constructs references based on the time of the knot points
+   * @brief construct references assuming initial com is the from the nominal
+   * stand
+   */
+  void Generate();
+
+  /*!
+   * @brief Set the nominal reference configuration q_ref
    * @param com initial location of the com
    */
-  void Build(const Eigen::Vector3d& com);
+  void SetNominalReferenceConfiguration(const Eigen::VectorXd& q_ref) {
+    q_ref_ = q_ref;
+  }
 
   drake::trajectories::PiecewisePolynomial<double> com_trajectory_;
   drake::trajectories::PiecewisePolynomial<double> q_trajectory_;
@@ -59,15 +73,13 @@ class KcmpcReferenceGenerator{
   drake::trajectories::PiecewisePolynomial<double> contact_traj_;
   drake::trajectories::PiecewisePolynomial<double> momentum_trajectory_;
 
-
  private:
   const drake::multibody::MultibodyPlant<double>& plant_;
+  drake::systems::Context<double>* context_;
   const std::vector<dairlib::multibody::WorldPointEvaluator<double>> contacts_;
-  Eigen::VectorXd nominal_stand_;
+  Eigen::VectorXd q_ref_;
   KnotPoints<Eigen::Vector3d> com_vel_knot_points_;
   KnotPoints<Gait> gait_knot_points_;
   double m_;
   Eigen::Vector3d p_ScmBase_W_;
-  std::unique_ptr<drake::systems::Context<double>> context_;
-
 };
