+++ systems/trajectory_optimization/kinematic_centroidal_planner/kinematic_centroidal_constraints.h
