+++ systems/trajectory_optimization/kinematic_centroidal_planner/kinematic_centroidal_gains.h
@@ -1,8 +1,9 @@
 #pragma once
-#include "drake/common/yaml/yaml_read_archive.h"
 #include "yaml-cpp/yaml.h"
 
-struct KinematicCentroidalGains{
+#include "drake/common/yaml/yaml_read_archive.h"
+
+struct KinematicCentroidalGains {
   Eigen::Vector3d com_position;
   std::unordered_map<std::string, double> generalized_positions;
   std::unordered_map<std::string, double> generalized_velocities;
@@ -12,7 +13,7 @@ struct KinematicCentroidalGains{
   Eigen::Vector3d lin_momentum;
   Eigen::Vector3d ang_momentum;
 
-  double swing_foot_minimum_height;
+  double tol;
   template <typename Archive>
   void Serialize(Archive* a) {
     a->Visit(DRAKE_NVP(com_position));
@@ -23,6 +24,6 @@ struct KinematicCentroidalGains{
     a->Visit(DRAKE_NVP(contact_force));
     a->Visit(DRAKE_NVP(lin_momentum));
     a->Visit(DRAKE_NVP(ang_momentum));
-    a->Visit(DRAKE_NVP(swing_foot_minimum_height));
+    a->Visit(DRAKE_NVP(tol));
   }
 };
