+++ systems/trajectory_optimization/kinematic_centroidal_planner/reference_generation_utils.cc
@@ -1,7 +1,45 @@
-#include <iostream>
-#include "reference_generation_utils.h"
+#include "systems/trajectory_optimization/kinematic_centroidal_planner/reference_generation_utils.h"
 
 #include "multibody/multibody_utils.h"
+#include "examples/Cassie/cassie_utils.h"
+
+
+std::vector<dairlib::multibody::WorldPointEvaluator<double>> CreateContactPoints(const drake::multibody::MultibodyPlant<
+    double> &plant,
+                                                                                                                  double mu) {
+  auto left_toe_pair = dairlib::LeftToeFront(plant);
+  auto left_heel_pair = dairlib::LeftToeRear(plant);
+  auto right_toe_pair = dairlib::RightToeFront(plant);
+  auto right_heel_pair = dairlib::RightToeRear(plant);
+
+  std::vector<int> active_inds{0, 1, 2};
+
+  auto left_toe_eval = dairlib::multibody::WorldPointEvaluator<double>(
+      plant, left_toe_pair.first, left_toe_pair.second,
+      Eigen::Matrix3d::Identity(), Eigen::Vector3d::Zero(), active_inds);
+  left_toe_eval.set_frictional();
+  left_toe_eval.set_mu(mu);
+
+  auto left_heel_eval = dairlib::multibody::WorldPointEvaluator<double>(
+      plant, left_heel_pair.first, left_heel_pair.second,
+      Eigen::Matrix3d::Identity(), Eigen::Vector3d::Zero(), active_inds);
+  left_heel_eval.set_frictional();
+  left_heel_eval.set_mu(mu);
+
+  auto right_toe_eval = dairlib::multibody::WorldPointEvaluator<double>(
+      plant, right_toe_pair.first, right_toe_pair.second,
+      Eigen::Matrix3d::Identity(), Eigen::Vector3d::Zero(), active_inds);
+  right_toe_eval.set_frictional();
+  right_toe_eval.set_mu(mu);
+
+  auto right_heel_eval = dairlib::multibody::WorldPointEvaluator<double>(
+      plant, right_heel_pair.first, right_heel_pair.second,
+      Eigen::Matrix3d::Identity(), Eigen::Vector3d::Zero(), active_inds);
+  right_heel_eval.set_frictional();
+  right_heel_eval.set_mu(mu);
+
+  return {left_toe_eval, left_heel_eval, right_toe_eval, right_heel_eval};
+}
 
 drake::trajectories::PiecewisePolynomial<double> GenerateComTrajectory(
     const Eigen::Vector3d& current_com,
