+++ systems/trajectory_optimization/lcm_trajectory_systems.h
@@ -32,6 +32,9 @@ class LcmTrajectoryReceiver : public drake::systems::LeafSystem<double> {
   drake::systems::InputPortIndex trajectory_input_port_;
   drake::systems::OutputPortIndex trajectory_output_port_;
   const std::string trajectory_name_;
+
+  mutable LcmTrajectory lcm_traj_;
+  std::string nominal_stand_path_ = "examples/Cassie/saved_trajectories/kcmpc_stand";
 };
 
 }  // namespace systems
